Refactoring welcome window
