+++ src/main/java/com/java_academy/gui/WelcomeWindowController.java
@@ -1,21 +1,23 @@
 package com.java_academy.gui;
 
 
-import com.java_academy.Main;
-import com.java_academy.ServerApp;
+import com.java_academy.logic.Game;
 import com.java_academy.logic.model.MessageObject;
 import com.java_academy.logic.tools.BSLog;
-import com.java_academy.logic.tools.I18NResolver;
 import com.java_academy.network.Connector;
-import javafx.application.Platform;
+import com.java_academy.network.socket_provider.ServerSocketProvider;
+import com.java_academy.network.socket_provider.core.SocketProvider;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.TextField;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
 import java.io.IOException;
+import java.net.InetSocketAddress;
+import java.net.ServerSocket;
 
 /**
  * @author Bartlomiej Janik
@@ -25,13 +27,15 @@ public class WelcomeWindowController {
 
 	@FXML
     AnchorPane anchor;
+    @FXML
+    TextField ipTextField;
 	
     public void connectToTheServer() throws IOException {
         showGui();
     }
 
     public void createServerAndConnectClient()  {
-        ServerApp serverApp = new ServerApp();
+        Server serverApp = new Server();
         Thread serverThread = new Thread(serverApp);
         serverThread.start();
 
@@ -42,6 +46,29 @@ public class WelcomeWindowController {
         }
 
         showGui();
+    }
+
+    private class Server implements Runnable {
+        public Connector connector;
+
+        @Override
+        public void run() {
+            try {
+                ServerSocket serverSocket = new ServerSocket();
+                SocketProvider socketProvider = new ServerSocketProvider(serverSocket);
+                InetSocketAddress inetSocketAddress = new InetSocketAddress(ipTextField.getText(), 4000);
+                connector = new Connector(socketProvider);
+                BSLog.info(BSLog.getLogger(getClass()), "Server is up and running");
+                Game game = new Game(connector::sendMessage, () -> connector.closeConnection());
+                connector.addMessageReceiverListenerToSocketProvider(game);
+                if (connector.connect(inetSocketAddress)) {
+                    game.startGame();
+
+                }
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
     }
 
     private void showGui()  {
