+++ src/main/resources/WelcomeWindow.fxml
@@ -12,5 +12,9 @@
                 text="Create Game" onAction="#createServerAndConnectClient"/>
         <Button layoutX="372.0" layoutY="151.0" mnemonicParsing="false" prefHeight="191.0" prefWidth="228.0"
                 text="Connect to game" onAction="#connectToTheServer"/>
+
+        <TextField fx:id="ipTextField" layoutX="105.0" layoutY="7.0" prefHeight="31.0" prefWidth="174.0">
+        </TextField>
+        <Label fx:id="ipLabel" layoutX="14.0" layoutY="12.0" text="IP address"/>
     </children>
 </AnchorPane>
